move fileReader class into its file
